requirements.txt
pyserial
pillow
packbits
pdf2image
qrcode